requirements.txt.txt
streamlit==1.38.0
psycopg2-binary==2.9.9
pandas==2.2.2
plotly==5.24.1
python-dotenv==1.0.1
sqlalchemy==2.0.35
